% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cragg.R
\name{cragg_donald}
\alias{cragg_donald}
\title{Calculate the Cragg-Donald statistic for a given model.}
\usage{
cragg_donald(X, Y, Z, data = data.frame())
}
\arguments{
\item{X}{(formula). A one-sided formula of control variables.}

\item{Y}{(formula). A one-sided formula of endoenous variables (treatments)}

\item{Z}{(formula). A one-sided formula of instruments}

\item{data}{(dataframe). An optional dataframe, list, or environment
containing the variables used in the model. As with many of the base R
functions, if the variables are not found here, they may be searched for in
the environment cragg_donald() was called.}
}
\value{
(cd_test) results object of class "cd_test"
}
\description{
Calculate the Cragg-Donald statistic for a given model.
}
