% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cragg.R
\name{cragg_donald}
\alias{cragg_donald}
\title{Calculate the Cragg-Donald statistic for a given model.}
\usage{
cragg_donald(X, Y, Z, data_fr)
}
\arguments{
\item{X}{formula of control variables}

\item{Y}{fromula of treatment variable}

\item{Z}{formula of instrumental variables}
}
\value{
The Cragg-Donald Statistic from the specified model
}
\description{
Calculate the Cragg-Donald statistic for a given model.
}
