% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_and_yogo.R
\name{stock_yogo_test}
\alias{stock_yogo_test}
\title{Perform the Stock and Yogo test for weak instruments}
\usage{
stock_yogo_test(X, Y, Z, data, B = 0.05, size_bias = "bias")
}
\arguments{
\item{X}{(formula). A one-sided formula of control variables.}

\item{Y}{(formula). A one-sided formula of endoenous variables (treatments)}

\item{Z}{(formula). A one-sided formula of instruments}

\item{data}{(dataframe). An optional dataframe, list, or environment
containing the variables used in the model. As with many of the base R
functions, if the variables are not found here, they may be searched for in
the environment cragg_donald() was called.}

\item{B}{One of [.05, .1, .15, .2, .25, .3]. The maximum size of allowable bias relative}

\item{size_bias}{Either "bias" or "size". Whether to use a critical
value based on the maximum allowable bias relative to regular OLS (bias), or maximum
Wald test size distortion.}
}
\value{
(float) the recommended critical value.
}
\description{
Perform the Stock and Yogo test for weak instruments
}
